refactor(756): extract Jinja2 templates + modules from build.py
- Extract CSS_BASE → templates/dep-graph.css (579 lines)
- Extract THEME_SCRIPT → templates/theme.js
- Create Jinja2 templates: base.html, card.html, milestone_row.html, lane_col.html
- Extract modules: status.py, render.py, milestone.py, flatten.py, anchors.py, lane.py, sections.py
- Refactor build.py to use Jinja2 Environment (226 lines, under 300 target)
- Add jinja2 dependency to dep-graph pyproject.toml
- Add golden output verification test
- Remove obsolete test_build.py (tested removed _prepare_render_data)
