+++ scripts/dep-graph/dep_graph/build.py
@@ -22,7 +22,6 @@ Size: read from gh.json entry 'size' field (derived from size:* label in fetch),
 
 from __future__ import annotations
 
-import copy
 import json
 import shutil
 import sys
@@ -30,40 +29,30 @@ from dataclasses import dataclass
 from html import escape
 from pathlib import Path
 
-from .derive import derive_lane, derive_standalone_order, is_auto_derived_standalone
-from .keys import format_key, parse_key, repo_slug
+from jinja2 import Environment, PackageLoader
+
+from .milestone import _prepare_milestone_rows, render_milestone_row
+from .render import (
+    CardContext,
+    _render_missing_card,
+    _render_repo_badge,
+    derive_status,
+    display_title,
+    render_deps,
+)
 from .schema import LayoutValidationError, validate_layout
-from .titles import normalize_title
+from .sections import render_cross_deps
 
 # ---------------------------------------------------------------------------
-# Context dataclasses
+# Jinja2 environment
 # ---------------------------------------------------------------------------
 
-
-@dataclass(frozen=True, slots=True)
-class CardContext:
-    repo: str
-    issue_num: int
-    lane_code: str
-    lane_of: dict[tuple[str, int], str]
-    ovr: dict
-    gh_entry: dict | None
-    extra_blocked_by: list[tuple[str, int]]
-    extra_blocking: list[tuple[str, int]]
-    gh_issues: dict
-    title_rules: list[dict] | None
-    primary_repo: str = ""
+_jinja_env = Environment(loader=PackageLoader("dep_graph", "templates"))
 
 
-@dataclass(frozen=True, slots=True)
-class FlatLaneContext:
-    fl: dict
-    lane_of: dict[tuple[str, int], str]
-    gh_issues: dict
-    overrides: dict
-    extra_deps: dict
-    title_rules: list[dict] | None
-    primary_repo: str = ""
+# ---------------------------------------------------------------------------
+# Build paths dataclass
+# ---------------------------------------------------------------------------
 
 
 @dataclass(frozen=True, slots=True)
@@ -75,1267 +64,8 @@ class BuildPaths:
 
 
 # ---------------------------------------------------------------------------
-# Status derivation
+# Meta escaping
 # ---------------------------------------------------------------------------
-
-
-def _has_active_blockers(
-    gh_entry: dict,
-    extra_blocked_by: list[tuple[str, int]],
-    gh_issues: dict,
-    own_repo: str,
-) -> bool:
-    """Return True if any blocker is still open."""
-    for item in gh_entry.get("blocked_by", []):
-        if isinstance(item, dict):
-            key = format_key(item["repo"], item["issue"])
-        else:
-            key = format_key(own_repo, item) if own_repo else str(item)
-        if gh_issues.get(key, {}).get("state") != "closed":
-            return True
-    for dep_repo, dep_num in extra_blocked_by:
-        key = format_key(dep_repo, dep_num)
-        if gh_issues.get(key, {}).get("state") != "closed":
-            return True
-    return False
-
-
-def derive_status(
-    ovr: dict,
-    gh_entry: dict | None,
-    extra_blocked_by: list[tuple[str, int]],
-    gh_issues: dict,
-    repo: str = "",
-) -> str:
-    if "status" in ovr:
-        return ovr["status"]
-    if gh_entry is None:
-        return "ready"
-    if gh_entry.get("defer"):
-        return "defer"
-    if gh_entry.get("state") == "closed":
-        return "done"
-    if _has_active_blockers(gh_entry, extra_blocked_by, gh_issues, repo):
-        return "blocked"
-    return "ready"
-
-
-# ---------------------------------------------------------------------------
-# Deps rendering
-# ---------------------------------------------------------------------------
-
-
-def _ref_to_tuple(item: dict | int, own_repo: str) -> tuple[str, int]:
-    """Normalise a blocked_by/blocking item to (repo, issue_num)."""
-    if isinstance(item, dict):
-        return item["repo"], item["issue"]
-    # legacy plain int — belongs to card's own repo
-    return own_repo, int(item)
-
-
-def _collect_dep_lists(
-    gh_entry: dict | None,
-    extra_blocked_by: list[tuple[str, int]],
-    extra_blocking: list[tuple[str, int]],
-    own_repo: str = "",
-) -> tuple[list[tuple[str, int]], list[tuple[str, int]]]:
-    """Merge GH deps with extra deps.
-
-    Returns (blocked_by, blocking) as repo/issue tuples.
-    """
-    if gh_entry is None:
-        return list(extra_blocked_by), list(extra_blocking)
-
-    raw_bb = gh_entry.get("blocked_by", [])
-    blocked_by: list[tuple[str, int]] = [_ref_to_tuple(x, own_repo) for x in raw_bb]
-    seen_bb: set[tuple[str, int]] = set(blocked_by)
-    for t in extra_blocked_by:
-        if t not in seen_bb:
-            blocked_by.append(t)
-
-    raw_bl = gh_entry.get("blocking", [])
-    blocking: list[tuple[str, int]] = [_ref_to_tuple(x, own_repo) for x in raw_bl]
-    seen_bl: set[tuple[str, int]] = set(blocking)
-    for t in extra_blocking:
-        if t not in seen_bl:
-            blocking.append(t)
-
-    return blocked_by, blocking
-
-
-def _format_dep_parts(
-    blocked_by: list[tuple[str, int]],
-    blocking: list[tuple[str, int]],
-    lane_code: str,
-    lane_of: dict[tuple[str, int], str],
-    own_repo: str = "",
-) -> tuple[list[str], list[str]]:
-    """Split deps into plain (same-lane) and ext (cross-lane) parts."""
-    plain_parts: list[str] = []
-    ext_parts: list[str] = []
-
-    for ref_repo, n in blocked_by:
-        dep_lane = lane_of.get((ref_repo, n))
-        is_foreign = ref_repo != own_repo
-        if dep_lane == lane_code and not is_foreign:
-            plain_parts.append(f"\u2190#{n}")
-        elif is_foreign:
-            # Show owner/repo#N for cross-repo deps
-            ext_parts.append(f"\u2190{ref_repo}#{n}")
-        else:
-            ext_parts.append(f"\u2190{dep_lane.upper() if dep_lane else '?'}:#{n}")
-
-    for ref_repo, n in blocking:
-        dep_lane = lane_of.get((ref_repo, n))
-        is_foreign = ref_repo != own_repo
-        if dep_lane == lane_code and not is_foreign:
-            plain_parts.append(f"\u2192#{n}")
-        elif is_foreign:
-            ext_parts.append(f"\u2192{ref_repo}#{n}")
-        else:
-            ext_parts.append(f"\u2192{dep_lane.upper() if dep_lane else '?'}:#{n}")
-
-    return plain_parts, ext_parts
-
-
-def render_deps(ctx: CardContext) -> str:
-    extra_deps_ext: list[str] = ctx.ovr.get("extra_deps_ext", [])
-
-    blocked_by, blocking = _collect_dep_lists(
-        ctx.gh_entry, ctx.extra_blocked_by, ctx.extra_blocking, ctx.repo
-    )
-    plain_parts, ext_parts = _format_dep_parts(
-        blocked_by, blocking, ctx.lane_code, ctx.lane_of, ctx.repo
-    )
-    ext_parts = ext_parts + extra_deps_ext
-
-    if not plain_parts and not ext_parts:
-        return '<span class="none">no deps</span>'
-
-    result = escape(" ".join(plain_parts)) if plain_parts else ""
-    if ext_parts:
-        ext_html = f'<span class="ext">{" ".join(escape(x) for x in ext_parts)}</span>'
-        result = (result + " " + ext_html).strip()
-    return result
-
-
-# ---------------------------------------------------------------------------
-# Card HTML
-# ---------------------------------------------------------------------------
-
-
-def display_title(
-    issue_num: int,
-    ovr: dict,
-    gh_entry: dict | None,
-    title_rules: list[dict] | None,
-) -> str:
-    # Per-issue override wins
-    if "title" in ovr:
-        return escape(ovr["title"])
-    if gh_entry:
-        raw = gh_entry["title"]
-        normalized = normalize_title(raw, title_rules)
-        return escape(normalized if normalized else raw)
-    return f"#{issue_num}"
-
-
-def _render_repo_badge(repo: str, primary_repo: str) -> str:
-    """Return HTML for a repo badge on foreign cards, or empty string if native."""
-    if repo == primary_repo:
-        return ""
-    name = repo.split("/", 1)[1] if "/" in repo else repo
-    # data-repo-badge carries the marker string; CSS targets .rbadge.
-    return (
-        f'<span class="rbadge" data-repo-badge title="{escape(repo)}">'
-        f"{escape(name)}</span>"
-    )
-
-
-def _render_missing_card(ref_repo: str, issue_num: int, anchor_attr: str = "") -> str:
-    """Return placeholder HTML for an IssueRef absent from gh.json."""
-    key = f"{ref_repo}#{issue_num}"
-    print(f"WARN: missing {key} in gh.json", file=sys.stderr)
-    card_id = f"card-{repo_slug(ref_repo)}-{issue_num}"
-    missing_attrs = f' id="{card_id}"{anchor_attr} data-missing="true"'
-    return (
-        f'<div class="card card--missing"{missing_attrs}>'
-        f'<span class="card-missing-label">not-found</span>'
-        f'<span class="card-issue">#{issue_num}</span>'
-        f'<span class="card-missing-repo">{escape(ref_repo)}</span>'
-        f"</div>"
-    )
-
-
-def render_card(ctx: CardContext, anchor_attr: str = "") -> str:
-    if ctx.gh_entry is None and ctx.repo:
-        return _render_missing_card(ctx.repo, ctx.issue_num, anchor_attr)
-    status = derive_status(
-        ctx.ovr,
-        ctx.gh_entry,
-        ctx.extra_blocked_by,
-        ctx.gh_issues,
-        ctx.repo,
-    )
-    # size: override wins, then gh.json 'size' field (from size:* label), then empty
-    size = (
-        ctx.ovr.get("size") or (ctx.gh_entry.get("size") if ctx.gh_entry else "") or ""
-    )
-    title = display_title(ctx.issue_num, ctx.ovr, ctx.gh_entry, ctx.title_rules)
-    size_html = f'<span class="size">{escape(size)}</span>' if size else ""
-    deps_html = render_deps(ctx)
-    repo_badge_html = _render_repo_badge(ctx.repo, ctx.primary_repo)
-    card_id = f"card-{repo_slug(ctx.repo)}-{ctx.issue_num}"
-    top_inner = f'<span class="num">#{ctx.issue_num}</span>{repo_badge_html}{size_html}'
-    return (
-        f'<div class="card {status}" id="{card_id}"{anchor_attr}>'
-        f'<div class="top">{top_inner}</div>'
-        f'<div class="title">{title}</div>'
-        f'<div class="deps">{deps_html}</div>'
-        f"</div>"
-    )
-
-
-# ---------------------------------------------------------------------------
-# Anchor / spacer computation
-# ---------------------------------------------------------------------------
-
-
-def compute_slot_index(flat_rows: list[dict], target_issue: int) -> int:
-    """1-based slot index of target_issue in a flat row list (bands + cards)."""
-    idx = 0
-    for row in flat_rows:
-        if row.get("band"):
-            idx += 1
-        elif "issue" in row:
-            idx += 1
-            if row["issue"] == target_issue:
-                return idx
-    return -1
-
-
-def _collect_anchors(flat_lanes: list[dict]) -> dict[str, tuple[str, int]]:
-    """Pass 1: collect anchor positions from all lanes."""
-    anchors: dict[str, tuple[str, int]] = {}
-    for fl in flat_lanes:
-        code = fl["code"]
-        for row in fl["flat_rows"]:
-            if "issue" not in row or "anchor" not in row:
-                continue
-            slot = compute_slot_index(fl["flat_rows"], row["issue"])
-            if slot == -1:
-                print(
-                    f"WARN: anchor '{row['anchor']}' issue #{row['issue']} "
-                    f"not found in lane {code}",
-                    file=sys.stderr,
-                )
-                continue
-            anchors[row["anchor"]] = (code, slot)
-    return anchors
-
-
-def _compute_lane_insertions(
-    fl: dict,
-    anchors: dict[str, tuple[str, int]],
-) -> list[tuple[int, int]]:
-    """Compute (insert_before_index, count) pairs for one lane."""
-    code = fl["code"]
-    rows = fl["flat_rows"]
-    insertions: list[tuple[int, int]] = []
-
-    for i, row in enumerate(rows):
-        if "issue" not in row or "anchor_after" not in row:
-            continue
-        anchor_id = row["anchor_after"]
-        if anchor_id not in anchors:
-            print(
-                f"WARN: anchor_after '{anchor_id}' in lane {code}: unknown anchor",
-                file=sys.stderr,
-            )
-            continue
-        _, ref_slot = anchors[anchor_id]
-        target_slot = compute_slot_index(rows, row["issue"])
-        if target_slot == -1:
-            continue
-
-        pg = row.get("par_group")
-        insert_before = i
-        if pg is not None:
-            for j in range(i - 1, -1, -1):
-                if rows[j].get("par_group") == pg:
-                    insert_before = j
-                else:
-                    break
-
-        needed = (ref_slot + 1) - target_slot
-        if needed <= 0:
-            print(
-                f"WARN: anchor_after '{anchor_id}' lane {code}: "
-                f"slot {target_slot} >= {ref_slot}+1",
-                file=sys.stderr,
-            )
-            continue
-        insertions.append((insert_before, needed))
-    return insertions
-
-
-def inject_spacers(flat_lanes: list[dict]) -> list[dict]:
-    """Insert synthetic spacer rows for anchor-based cross-lane alignment."""
-    flat_lanes = copy.deepcopy(flat_lanes)
-
-    anchors = _collect_anchors(flat_lanes)
-    if not anchors:
-        return flat_lanes
-
-    # Pass 2: insert spacers
-    for fl in flat_lanes:
-        rows = fl["flat_rows"]
-        insertions = _compute_lane_insertions(fl, anchors)
-        for insert_before, count in sorted(
-            insertions, key=lambda x: x[0], reverse=True
-        ):
-            rows[insert_before:insert_before] = [{"spacer": True}] * count
-
-    return flat_lanes
-
-
-# ---------------------------------------------------------------------------
-# Lane flattening (new schema → flat_rows)
-# ---------------------------------------------------------------------------
-
-
-def _check_drift(repo: str, n: int, code: str, gh_issues: dict) -> None:
-    """Emit stderr warning if GH lane label differs from layout lane."""
-    gh_entry = gh_issues.get(format_key(repo, n))
-    if not gh_entry:
-        return
-    gh_lane = gh_entry.get("lane_label")
-    if gh_lane and gh_lane != code:
-        print(
-            f"WARN drift: layout says {format_key(repo, n)} → lane {code},"
-            f" gh label says {gh_lane}",
-            file=sys.stderr,
-        )
-
-
-def _build_issue_to_pg(pg_map: dict) -> dict[tuple[str, int], str]:
-    """Build (repo, issue) → par_group_id map from lane par_groups."""
-    issue_to_pg: dict[tuple[str, int], str] = {}
-    for gid, members in pg_map.items():
-        for ref in members:
-            if isinstance(ref, dict):
-                issue_to_pg[(ref["repo"], ref["issue"])] = gid
-            else:
-                issue_to_pg[("", int(ref))] = gid
-    return issue_to_pg
-
-
-def _build_band_before(bands: list) -> dict[tuple[str, int] | int, str]:
-    """Build ref → band text map from lane bands list."""
-    band_before: dict[tuple[str, int] | int, str] = {}
-    for b in bands:
-        bef = b["before"]
-        if isinstance(bef, dict):
-            band_before[(bef["repo"], bef["issue"])] = b["text"]
-        else:
-            band_before[int(bef)] = b["text"]
-    return band_before
-
-
-def _flatten_order_item(
-    item: dict | int,
-    overrides: dict,
-    issue_to_pg: dict[tuple[str, int], str],
-    band_before: dict[tuple[str, int] | int, str],
-) -> tuple[dict | None, dict]:
-    """Convert one order item to an optional band row + an issue row."""
-    if isinstance(item, dict):
-        repo: str = item["repo"]
-        n: int = item["issue"]
-        ref_key: tuple[str, int] | int = (repo, n)
-    else:
-        repo = ""
-        n = int(item)
-        ref_key = n
-
-    band_row: dict | None = None
-    if ref_key in band_before:
-        band_row = {"band": band_before[ref_key]}
-
-    ovr_key = f"{repo}#{n}" if repo else str(n)
-    ovr = overrides.get(ovr_key, {})
-    row: dict = {"issue": n, "repo": repo}
-    pg = issue_to_pg.get((repo, n)) or issue_to_pg.get(("", n))
-    if pg is not None:
-        row["par_group"] = pg
-    if "anchor" in ovr:
-        row["anchor"] = ovr["anchor"]
-    if "anchor_after" in ovr:
-        row["anchor_after"] = ovr["anchor_after"]
-    return band_row, row
-
-
-def flatten_lane(
-    lane: dict,
-    overrides: dict,
-    label_drift_check: bool,
-    gh_issues: dict,
-) -> dict:
-    """Convert new-schema lane into flat_rows consumed by render/inject."""
-    code = lane["code"]
-    order = lane.get("order", [])
-    issue_to_pg = _build_issue_to_pg(lane.get("par_groups", {}))
-    band_before = _build_band_before(lane.get("bands", []))
-
-    flat_rows: list[dict] = []
-    for item in order:
-        band_row, row = _flatten_order_item(item, overrides, issue_to_pg, band_before)
-        if band_row is not None:
-            flat_rows.append(band_row)
-        flat_rows.append(row)
-        if label_drift_check and row.get("repo"):
-            _check_drift(row["repo"], row["issue"], code, gh_issues)
-
-    return {
-        "code": code,
-        "name": lane["name"],
-        "color": lane["color"],
-        "epic": lane.get("epic"),
-        "flat_rows": flat_rows,
-    }
-
-
-# ---------------------------------------------------------------------------
-# Lane HTML rendering
-# ---------------------------------------------------------------------------
-
-
-def _close_par(row_htmls: list[str], current_par: str | None) -> None:
-    """Close an open par-group div if one is active."""
-    if current_par is not None:
-        row_htmls.append("    </div>")
-
-
-def _render_issue_row(
-    row: dict,
-    ctx: FlatLaneContext,
-    row_htmls: list[str],
-    current_par: str | None,
-) -> str | None:
-    """Render an issue row. Returns updated current_par."""
-    code = ctx.fl["code"]
-    extra_blocked_by_map: dict[str, list[int]] = ctx.extra_deps.get(
-        "extra_blocked_by", {}
-    )
-    extra_blocking_map: dict[str, list[int]] = ctx.extra_deps.get("extra_blocking", {})
-
-    n = row["issue"]
-    row_repo: str = row.get("repo", "")
-    pg: str | None = row.get("par_group")
-    if pg != current_par:
-        _close_par(row_htmls, current_par)
-        current_par = None
-        if pg is not None:
-            row_htmls.append('    <div class="par">')
-            current_par = pg
-    ovr_key = format_key(row_repo, n) if row_repo else str(n)
-    ovr = ctx.overrides.get(ovr_key, {})
-    gh_key = format_key(row_repo, n) if row_repo else str(n)
-    gh_entry = ctx.gh_issues.get(gh_key)
-    if gh_entry is None and not row_repo:
-        # fallback for legacy int-keyed overrides
-        gh_entry = ctx.gh_issues.get(str(n))
-    bb_key = format_key(row_repo, n) if row_repo else str(n)
-    bl_key = format_key(row_repo, n) if row_repo else str(n)
-    _raw_bb = extra_blocked_by_map.get(bb_key, [])
-    _raw_bl = extra_blocking_map.get(bl_key, [])
-
-    def _as_ref(k: str | int) -> tuple[str, int]:
-        if isinstance(k, str):
-            return parse_key(k)
-        return (row_repo, int(k))
-
-    extra_bb: list[tuple[str, int]] = [_as_ref(k) for k in _raw_bb]
-    extra_bl: list[tuple[str, int]] = [_as_ref(k) for k in _raw_bl]
-    anchor_attr = ""
-    if "anchor" in row:
-        anchor_attr = f' data-anchor="{escape(row["anchor"])}"'
-    elif "anchor_after" in row:
-        anchor_attr = f' data-anchor-after="{escape(row["anchor_after"])}"'
-    card_ctx = CardContext(
-        repo=row_repo,
-        issue_num=n,
-        lane_code=code,
-        lane_of=ctx.lane_of,
-        ovr=ovr,
-        gh_entry=gh_entry,
-        extra_blocked_by=extra_bb,
-        extra_blocking=extra_bl,
-        gh_issues=ctx.gh_issues,
-        title_rules=ctx.title_rules,
-        primary_repo=ctx.primary_repo,
-    )
-    card = render_card(card_ctx, anchor_attr)
-    indent = "      " if current_par is not None else "    "
-    row_htmls.append(f"{indent}{card}")
-    return current_par
-
-
-def _render_row(
-    row: dict,
-    ctx: FlatLaneContext,
-    row_htmls: list[str],
-    current_par: str | None,
-) -> str | None:
-    """Render one row and append to row_htmls. Returns updated current_par."""
-    if row.get("spacer"):
-        _close_par(row_htmls, current_par)
-        current_par = None
-        row_htmls.append('    <div class="spacer" aria-hidden="true"></div>')
-    elif row.get("band"):
-        _close_par(row_htmls, current_par)
-        current_par = None
-        row_htmls.append(f'    <div class="ms-band">{escape(row["band"])}</div>')
-    elif "issue" in row:
-        current_par = _render_issue_row(row, ctx, row_htmls, current_par)
-    return current_par
-
-
-def render_flat_lane(ctx: FlatLaneContext) -> str:
-    fl = ctx.fl
-    code = fl["code"]
-    name = fl["name"]
-    color = fl["color"]
-    epic = fl.get("epic")
-    rows = fl["flat_rows"]
-
-    parts: list[str] = []
-    parts.append(f'  <div class="lane" data-lane="{escape(color)}">')
-    parts.append(
-        f'    <div class="lane-head">'
-        f'<span class="code">{escape(code.upper())}</span>'
-        f'<span class="name">{escape(name)}</span></div>'
-    )
-
-    row_htmls: list[str] = []
-    current_par: str | None = None
-
-    for row in rows:
-        current_par = _render_row(row, ctx, row_htmls, current_par)
-
-    # close any open par group
-    _close_par(row_htmls, current_par)
-
-    if epic:
-        defer_class = " defer" if epic.get("defer") else ""
-        epic_issue = epic["issue"]
-        epic_label = escape(epic["label"])
-        epic_tag = escape(epic["tag"])
-        parts.append(f'    <div class="epic-wrap{defer_class}">')
-        parts.append(
-            f'      <div class="epic-banner">'
-            f"<span>#{epic_issue} \u00b7 {epic_label}</span>"
-            f'<span class="tag">{epic_tag}</span></div>'
-        )
-        for rh in row_htmls:
-            parts.append("  " + rh)
-        parts.append("    </div>")
-    else:
-        parts.extend(row_htmls)
-
-    parts.append("  </div>")
-    return "\n".join(parts)
-
-
-# ---------------------------------------------------------------------------
-# Untriaged section
-# ---------------------------------------------------------------------------
-
-
-def render_untriaged(
-    untriaged: list[tuple[str, int]],
-    gh_issues: dict,
-    primary_repo: str = "",
-) -> str:
-    if not untriaged:
-        return ""
-    cards = []
-    for repo, n in untriaged:
-        gh_entry = gh_issues.get(format_key(repo, n))
-        title = escape(gh_entry.get("title", f"#{n}")) if gh_entry else f"#{n}"
-        cards.append(
-            f'    <div class="card ready">'
-            f'<div class="top"><span class="num">#{n}</span></div>'
-            f'<div class="title">{title}</div>'
-            f'<div class="deps"><span class="none">untriaged</span></div>'
-            f"</div>"
-        )
-    inner = "\n".join(cards)
-    return f"""\
-<div class="standalone">
-  <div class="standalone-head">
-    <span class="label">Untriaged</span>
-    <span class="hint">\u2014 labeled but not in any lane order</span>
-  </div>
-  <div class="standalone-grid">
-{inner}
-  </div>
-</div>
-
-"""
-
-
-# ---------------------------------------------------------------------------
-# Standalone section
-# ---------------------------------------------------------------------------
-
-
-def render_standalone(
-    order: list[tuple[str, int]],
-    gh_issues: dict,
-    overrides: dict,
-    title_rules: list[dict] | None,
-    primary_repo: str = "",
-) -> str:
-    cards = []
-    for repo, n in order:
-        ovr_key = format_key(repo, n) if repo else str(n)
-        ovr = overrides.get(ovr_key, {})
-        gh_entry = gh_issues.get(format_key(repo, n) if repo else str(n))
-        status = derive_status(ovr, gh_entry, [], gh_issues, repo)
-        size = ovr.get("size") or (gh_entry.get("size") if gh_entry else "") or ""
-        title = display_title(n, ovr, gh_entry, title_rules)
-        size_html = f'<span class="size">{escape(size)}</span>' if size else ""
-        cards.append(
-            f'    <div class="card {status}">'
-            f'<div class="top"><span class="num">#{n}</span>{size_html}</div>'
-            f'<div class="title">{title}</div>'
-            f'<div class="deps"><span class="none">no deps</span></div>'
-            f"</div>"
-        )
-    return "\n".join(cards)
-
-
-# ---------------------------------------------------------------------------
-# Cross-deps section
-# ---------------------------------------------------------------------------
-
-
-def render_cross_deps(cross_deps: list[dict]) -> str:
-    items = []
-    for cd in cross_deps:
-        kind = escape(cd["kind"])
-        text = escape(cd["text"])
-        items.append(f'    <li><span class="kind">{kind}</span>{text}</li>')
-    return "\n".join(items)
-
-
-# ---------------------------------------------------------------------------
-# CSS (parameterized — lane color vars injected from layout)
-# ---------------------------------------------------------------------------
-
-CSS_BASE = """\
-:root {
-  --slot-h: 56px;
-}
-:root, [data-theme="dark"] {
-  --bg:          #0d1117;
-  --bg-panel:    #13191f;
-  --bg-card:     #161b22;
-  --bg-cell:     #0f141a;
-  --bg-epic:     rgba(232,93,4,0.05);
-  --border:      #21262d;
-  --border-hi:   #30363d;
-  --text:        #fafafa;
-  --text-muted:  #8b93a1;
-  --text-dim:    #6b7280;
-  --accent:      #e85d04;
-  --accent-hi:   #f97316;
-  --lane-a1:     #22c55e;
-  --lane-a2:     #6366f1;
-  --lane-b:      #3b82f6;
-  --lane-c1:     #a855f7;
-  --lane-c2:     #d946ef;
-  --lane-c3:     #ec4899;
-  --lane-d:      #06b6d4;
-  --lane-e:      #eab308;
-  --lane-f:      #10b981;
-  --lane-g:      #f97316;
-  --lane-h:      #f59e0b;
-  --lane-i:      #14b8a6;
-  --lane-indep:  #94a3b8;
-  --status-done: #16a34a;
-  --status-ready:#facc15;
-  --status-blocked:#fb923c;
-  --status-defer:#64748b;
-  --arrow:       #4b5563;
-  --dep-in:      #94a3b8;
-  --dep-out:     #94a3b8;
-  --dep-ext:     #f97316;
-{extra_vars}
-[data-theme="light"] {
-  --bg:          #fafaf9;
-  --bg-panel:    #ffffff;
-  --bg-card:     #f4f4f0;
-  --bg-cell:     #ffffff;
-  --bg-epic:     rgba(232,93,4,0.04);
-  --border:      #d6d3d1;
-  --border-hi:   #a8a29e;
-  --text:        #1c1917;
-  --text-muted:  #57534e;
-  --text-dim:    #78716c;
-  --accent:      #c2410c;
-  --accent-hi:   #e85d04;
-  --arrow:       #a8a29e;
-  --dep-in:      #6b7280;
-}
-* { box-sizing: border-box; margin: 0; padding: 0; }
-body {
-  background: var(--bg);
-  color: var(--text);
-  font-family: 'Inter', system-ui, sans-serif;
-  padding: 24px;
-  min-height: 100vh;
-  font-size: 12px;
-  line-height: 1.4;
-}
-header {
-  max-width: 100%;
-  margin: 0 auto 18px;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  gap: 24px;
-  flex-wrap: wrap;
-}
-h1 {
-  font-size: 22px;
-  font-weight: 700;
-  letter-spacing: -0.02em;
-}
-h1 span { color: var(--accent); }
-.subtitle {
-  color: var(--text-muted);
-  font-size: 12px;
-  margin-top: 4px;
-}
-.legend {
-  display: flex;
-  gap: 6px;
-  flex-wrap: wrap;
-  font-size: 10px;
-  color: var(--text-muted);
-}
-.legend .pill {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  padding: 3px 7px;
-  background: var(--bg-card);
-  border: 1px solid var(--border);
-  border-radius: 4px;
-  font-family: 'JetBrains Mono', monospace;
-}
-.dot {
-  display: inline-block;
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-.dot.done    { background: var(--status-done); }
-.dot.ready   { background: var(--status-ready); }
-.dot.blocked { background: var(--status-blocked); }
-.dot.defer   { background: var(--status-defer); }
-
-/* Flex-column lanes layout */
-.lanes {
-  max-width: 100%;
-  margin: 0 auto;
-  display: flex;
-  gap: 8px;
-  align-items: stretch;
-  background: var(--bg-panel);
-  padding: 12px;
-  border: 1px solid var(--border);
-  border-radius: 10px;
-}
-.lane {
-  flex: 1 1 0;
-  min-width: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-.lane-head {
-  padding: 8px 6px;
-  text-align: center;
-  background: var(--bg-card);
-  border: 1px solid var(--border);
-  border-top: 3px solid;
-  border-radius: 5px 5px 0 0;
-}
-.lane-head .code {
-  font-family: 'JetBrains Mono', monospace;
-  font-weight: 700;
-  font-size: 13px;
-}
-.lane-head .name {
-  display: block;
-  font-size: 10px;
-  color: var(--text-muted);
-  margin-top: 2px;
-  font-weight: 500;
-}
-.lane[data-lane="a1"]    > .lane-head { border-top-color: var(--lane-a1); }
-.lane[data-lane="a2"]    > .lane-head { border-top-color: var(--lane-a2); }
-.lane[data-lane="b"]     > .lane-head { border-top-color: var(--lane-b); }
-.lane[data-lane="c1"]    > .lane-head { border-top-color: var(--lane-c1); }
-.lane[data-lane="c2"]    > .lane-head { border-top-color: var(--lane-c2); }
-.lane[data-lane="c3"]    > .lane-head { border-top-color: var(--lane-c3); }
-.lane[data-lane="d"]     > .lane-head { border-top-color: var(--lane-d); }
-.lane[data-lane="e"]     > .lane-head { border-top-color: var(--lane-e); }
-.lane[data-lane="f"]     > .lane-head { border-top-color: var(--lane-f); }
-.lane[data-lane="g"]     > .lane-head { border-top-color: var(--lane-g); }
-.lane[data-lane="h"]     > .lane-head { border-top-color: var(--lane-h); }
-.lane[data-lane="i"]     > .lane-head { border-top-color: var(--lane-i); }
-.lane[data-lane="indep"] > .lane-head { border-top-color: var(--lane-indep); }
-{extra_selectors}
-/* Epic wrap */
-.epic-wrap {
-  background: var(--bg-epic);
-  border: 1px solid var(--accent);
-  border-radius: 6px;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  padding: 0 5px 6px;
-  flex: 1 1 auto;
-}
-.epic-wrap.defer { opacity: 0.7; }
-.epic-banner {
-  background: var(--accent);
-  color: #fff;
-  padding: 3px 8px;
-  font-family: 'JetBrains Mono', monospace;
-  font-size: 11px;
-  font-weight: 700;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 6px;
-  border-radius: 5px 5px 0 0;
-  overflow: hidden;
-  white-space: nowrap;
-  margin: 0 -5px;
-}
-.epic-banner .tag {
-  font-size: 8px;
-  opacity: 0.85;
-  text-transform: uppercase;
-  letter-spacing: 0.1em;
-  flex-shrink: 0;
-}
-
-/* Section band (milestone label) */
-.ms-band {
-  display: flex;
-  align-items: center;
-  height: 18px;
-  line-height: 18px;
-  font-size: 9px;
-  font-family: 'JetBrains Mono', monospace;
-  color: var(--text-dim);
-  text-transform: uppercase;
-  letter-spacing: 0.1em;
-}
-.ms-band::after {
-  content: "";
-  flex: 1;
-  height: 1px;
-  background: var(--border);
-  margin-left: 6px;
-}
-
-/* Spacer — invisible slot placeholder */
-.spacer {
-  height: var(--slot-h);
-}
-
-/* Parallel siblings group */
-.par {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  border-left: 2px solid var(--text-dim);
-  padding-left: 6px;
-  margin-left: 2px;
-}
-
-/* Card — fixed height for slot uniformity */
-.card {
-  background: var(--bg-cell);
-  border: 1px solid var(--border);
-  border-left: 3px solid var(--status-defer);
-  border-radius: 4px;
-  padding: 5px 7px;
-  font-size: 11px;
-  min-width: 0;
-  height: var(--slot-h);
-  min-height: var(--slot-h);
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-  flex: 0 0 auto;
-}
-.card.done     { border-left-color: var(--status-done); opacity: 0.55; }
-.card.ready    { border-left-color: var(--status-ready); }
-.card.blocked  { border-left-color: var(--status-blocked); }
-.card.defer    { border-left-color: var(--status-defer); opacity: 0.65; }
-
-.card .top {
-  display: flex;
-  justify-content: space-between;
-  align-items: baseline;
-  gap: 4px;
-}
-.card .num {
-  font-family: 'JetBrains Mono', monospace;
-  font-weight: 700;
-  font-size: 11px;
-}
-.card.done .num::after { content: " \u2713"; color: var(--status-done); }
-.card .size {
-  font-family: 'JetBrains Mono', monospace;
-  font-size: 9px;
-  color: var(--text-dim);
-}
-.card .title {
-  font-size: 10px;
-  color: var(--text-muted);
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-.card .deps {
-  font-family: 'JetBrains Mono', monospace;
-  font-size: 9px;
-  display: flex;
-  flex-wrap: nowrap;
-  gap: 4px 6px;
-  margin-top: 2px;
-  color: var(--dep-in);
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-.card .deps .ext { color: var(--dep-ext); font-weight: 600; }
-.card .deps .none { color: var(--text-dim); opacity: 0.6; }
-
-.rbadge {
-  display: inline-block;
-  padding: 1px 6px;
-  margin-left: 6px;
-  font-size: 0.7em;
-  border-radius: 4px;
-  background: var(--border-hi, #30363d);
-  color: var(--text-dim, #6b7280);
-  font-weight: 500;
-}
-
-.card--missing {
-  border: 2px dashed var(--status-blocked, #fb923c);
-  background: rgba(251,146,60,0.06);
-  color: var(--text-muted, #8b93a1);
-  padding: 8px;
-  font-style: italic;
-}
-
-/* Milestone-first layout */
-.milestone-row {
-  max-width: 100%;
-  margin: 0 auto 12px;
-  background: var(--bg-panel);
-  border: 1px solid var(--border);
-  border-radius: 10px;
-  overflow: hidden;
-}
-.milestone-row-header {
-  padding: 10px 14px;
-  background: var(--bg-card);
-  border-bottom: 1px solid var(--border);
-  display: flex;
-  align-items: baseline;
-  gap: 10px;
-  font-family: 'JetBrains Mono', monospace;
-}
-.milestone-row-header[data-milestone="none"] {
-  background: linear-gradient(90deg, var(--bg-card) 0%, rgba(148,163,184,0.1) 100%);
-  border-left: 3px solid var(--lane-indep);
-}
-.ms-anchor {
-  font-size: 11px;
-  font-weight: 700;
-  color: var(--text-dim);
-  text-transform: uppercase;
-  letter-spacing: 0.1em;
-}
-.ms-label {
-  font-size: 13px;
-  font-weight: 700;
-  color: var(--accent);
-}
-.ms-hint {
-  font-size: 10px;
-  color: var(--text-dim);
-}
-.milestone-cards {
-  padding: 10px;
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(180px, 1fr));
-  gap: 6px;
-}
-.milestone-cols {
-  padding: 10px;
-  display: flex;
-  gap: 8px;
-  overflow-x: auto;
-}
-.lane-col {
-  flex: 0 0 auto;
-  min-width: 180px;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-.lane-col-head {
-  padding: 6px 8px;
-  text-align: center;
-  background: var(--bg-card);
-  border: 1px solid var(--border);
-  border-top: 3px solid;
-  border-radius: 5px 5px 0 0;
-  font-family: 'JetBrains Mono', monospace;
-}
-.lane-col-head .code {
-  font-weight: 700;
-  font-size: 13px;
-}
-.lane-col-head .name {
-  display: block;
-  font-size: 10px;
-  color: var(--text-muted);
-  margin-top: 2px;
-}
-.lane-col-head[data-lane="a1"] { border-top-color: var(--lane-a1); }
-.lane-col-head[data-lane="b"]  { border-top-color: var(--lane-b); }
-.lane-col-head[data-lane="c1"] { border-top-color: var(--lane-c1); }
-.lane-col-head[data-lane="d"]  { border-top-color: var(--lane-d); }
-.lane-col-head[data-lane="e"]  { border-top-color: var(--lane-e); }
-.lane-col-head[data-lane="f"]  { border-top-color: var(--lane-f); }
-.lane-col-head[data-lane="g"]  { border-top-color: var(--lane-g); }
-.lane-col-head[data-lane="h"]  { border-top-color: var(--lane-h); }
-.lane-col-head[data-lane="i"]  { border-top-color: var(--lane-i); }
-.lane-col-head[data-lane="indep"] { border-top-color: var(--lane-indep); }
-.lane-col-cards {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-.lane-col--split {
-  flex-direction: column;
-}
-.lane-sub-cols {
-  display: flex;
-  gap: 4px;
-  flex: 1;
-}
-.lane-sub-col {
-  flex: 0 0 auto;
-  min-width: 160px;
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-}
-.lane-sub-head {
-  padding: 4px 6px;
-  text-align: center;
-  background: var(--bg-cell);
-  border-left: 3px solid;
-  border-radius: 3px;
-  font-family: 'JetBrains Mono', monospace;
-  font-size: 10px;
-}
-.lane-sub-head .code {
-  font-weight: 700;
-}
-.lane-sub-head[data-lane="a1"] { border-left-color: var(--lane-a1); }
-.lane-sub-head[data-lane="a2"] { border-left-color: var(--lane-a2); }
-.lane-sub-head[data-lane="c1"] { border-left-color: var(--lane-c1); }
-.lane-sub-head[data-lane="c2"] { border-left-color: var(--lane-c2); }
-.lane-sub-head[data-lane="c3"] { border-left-color: var(--lane-c3); }
-.lane-sub-cards {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-.mini-lane-head {
-  grid-column: 1 / -1;
-  padding: 6px 8px;
-  margin: 4px 0 2px;
-  background: var(--bg-cell);
-  border-left: 3px solid;
-  border-radius: 4px;
-  font-family: 'JetBrains Mono', monospace;
-  font-size: 10px;
-  display: flex;
-  gap: 8px;
-}
-.mini-lane-head .code {
-  font-weight: 700;
-}
-.mini-lane-head .name {
-  color: var(--text-muted);
-  font-weight: 500;
-}
-.mini-lane-head[data-lane="a1"] { border-left-color: var(--lane-a1); }
-.mini-lane-head[data-lane="a2"] { border-left-color: var(--lane-a2); }
-.mini-lane-head[data-lane="b"]  { border-left-color: var(--lane-b); }
-.mini-lane-head[data-lane="c1"] { border-left-color: var(--lane-c1); }
-.mini-lane-head[data-lane="c2"] { border-left-color: var(--lane-c2); }
-.mini-lane-head[data-lane="c3"] { border-left-color: var(--lane-c3); }
-.mini-lane-head[data-lane="d"]  { border-left-color: var(--lane-d); }
-.mini-lane-head[data-lane="e"]  { border-left-color: var(--lane-e); }
-.mini-lane-head[data-lane="f"]  { border-left-color: var(--lane-f); }
-.mini-lane-head[data-lane="g"]  { border-left-color: var(--lane-g); }
-.mini-lane-head[data-lane="h"]  { border-left-color: var(--lane-h); }
-.mini-lane-head[data-lane="i"]  { border-left-color: var(--lane-i); }
-
-.lane-badge {
-  display: inline-block;
-  padding: 1px 5px;
-  margin-left: 6px;
-  font-size: 0.7em;
-  border-radius: 4px;
-  background: var(--border);
-  color: var(--text-dim);
-  font-weight: 600;
-  font-family: 'JetBrains Mono', monospace;
-}
-
-.standalone {
-  max-width: 100%;
-  margin: 10px auto 0;
-  background: var(--bg-panel);
-  border: 1px solid var(--border);
-  border-radius: 10px;
-  overflow: hidden;
-}
-.standalone-head {
-  padding: 8px 14px;
-  background: var(--bg-card);
-  border-bottom: 1px solid var(--border);
-  display: flex;
-  align-items: baseline;
-  gap: 10px;
-  font-family: 'JetBrains Mono', monospace;
-}
-.standalone-head .label {
-  font-size: 12px;
-  font-weight: 700;
-  letter-spacing: 0.02em;
-  color: var(--text);
-}
-.standalone-head .hint {
-  font-size: 10px;
-  color: var(--text-muted);
-}
-.standalone-grid {
-  padding: 10px;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(160px, 1fr));
-  gap: 6px;
-}
-
-.cross-deps {
-  max-width: 100%;
-  margin: 18px auto 0;
-  padding: 14px 18px;
-  background: var(--bg-panel);
-  border: 1px solid var(--border);
-  border-radius: 10px;
-}
-.cross-deps h3 {
-  font-size: 13px;
-  font-weight: 600;
-  margin-bottom: 8px;
-}
-.cross-deps ul {
-  list-style: none;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
-  gap: 4px 18px;
-  font-family: 'JetBrains Mono', monospace;
-  font-size: 11px;
-  color: var(--text-muted);
-}
-.cross-deps li::before { content: "\u21b3 "; color: var(--accent); }
-.cross-deps .kind {
-  display: inline-block;
-  padding: 1px 5px;
-  border-radius: 2px;
-  font-size: 9px;
-  background: rgba(232,93,4,0.1);
-  color: var(--accent);
-  margin-right: 4px;
-}
-
-.theme-btn {
-  position: fixed;
-  top: 14px;
-  right: 14px;
-  padding: 6px 10px;
-  background: var(--bg-card);
-  border: 1px solid var(--border);
-  color: var(--text-muted);
-  font-family: inherit;
-  font-size: 11px;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-footer {
-  max-width: 100%;
-  margin: 16px auto 0;
-  text-align: center;
-  font-size: 11px;
-  color: var(--text-dim);
-  font-family: 'JetBrains Mono', monospace;
-}
-footer a { color: var(--accent); text-decoration: none; }
-"""
-
-THEME_SCRIPT = """\
-<script>
-(function() {
-  const btn = document.getElementById('theme-toggle');
-  const KEY = 'lyra-v2-graph-theme';
-  const saved = localStorage.getItem(KEY);
-  if (saved) document.documentElement.setAttribute('data-theme', saved);
-  function update() {
-    const cur = document.documentElement.getAttribute('data-theme');
-    btn.textContent = cur === 'dark' ? '\u25d1 light' : '\u25d0 dark';
-  }
-  update();
-  btn.addEventListener('click', () => {
-    const cur = document.documentElement.getAttribute('data-theme');
-    const next = cur === 'dark' ? 'light' : 'dark';
-    document.documentElement.setAttribute('data-theme', next);
-    localStorage.setItem(KEY, next);
-    update();
-  });
-})();
-</script>"""
 
 
 def _escape_meta(meta: dict, primary_repo: str) -> dict[str, str]:
@@ -1359,443 +89,8 @@ def _escape_meta(meta: dict, primary_repo: str) -> dict[str, str]:
 
 
 # ---------------------------------------------------------------------------
-# Milestone-first grouping (new architecture)
+# Main HTML builder
 # ---------------------------------------------------------------------------
-
-# Milestone display order: NO MILESTONE first, then M0-M5
-MILESTONE_ORDER = [
-    None,  # "No milestone" anchor at top
-    "M0 — NATS hardening",
-    "M1 — NATS maturity + containerize",
-    "M2 — LLM stack modernization",
-    "M3 — Observability",
-    "M4 — Hub statelessness",
-    "M5 — Plugin layer",
-]
-
-# Lane groups for column display (merge sub-lanes into major lanes)
-LANE_GROUPS = {
-    "A": ["a1", "a2"],  # NATS maturity + SDK
-    "B": ["b"],  # Containerize
-    "C": ["c1", "c2", "c3"],  # LLM stack (LiteLLM, harness, CLI)
-    "D": ["d"],  # Observability
-    "E": ["e"],  # Hub stateless
-    "F": ["f"],  # Plugins
-    "G": ["g"],  # Voice
-    "H": ["h"],  # Deploy ops
-    "I": ["i"],  # Vault ingest
-}
-
-# Lane group metadata
-LANE_GROUP_META = {
-    "A": {"name": "NATS", "color": "a1"},
-    "B": {"name": "Containerize", "color": "b"},
-    "C": {"name": "LLM Stack", "color": "c1"},
-    "D": {"name": "Observability", "color": "d"},
-    "E": {"name": "Stateless", "color": "e"},
-    "F": {"name": "Plugins", "color": "f"},
-    "G": {"name": "Voice", "color": "g"},
-    "H": {"name": "Deploy", "color": "h"},
-    "I": {"name": "Vault", "color": "i"},
-}
-
-
-def _lane_code_to_group(lane_code: str | None) -> str | None:
-    """Map a lane code (a1, a2, c1, etc.) to its group (A, C, etc.)."""
-    if lane_code is None:
-        return None
-    for group, codes in LANE_GROUPS.items():
-        if lane_code in codes:
-            return group
-    return None
-
-
-def _group_issues_by_milestone(
-    gh_issues: dict,
-    primary_repo: str,
-) -> dict[str | None, list[tuple[str, int, str | None, str | None]]]:
-    """Group issues by milestone.
-
-    Returns milestone → [(repo, issue_num, lane_code, lane_group)].
-    Issues with no milestone get key None. Lane_code is from lane_label field.
-    Lane_group is the major lane group (A, B, C, etc.).
-    """
-    groups: dict[str | None, list[tuple[str, int, str | None, str | None]]] = {}
-    for _key, entry in gh_issues.items():
-        if not entry:
-            continue
-        repo = entry.get("repo", primary_repo)
-        num = entry.get("number")
-        if num is None:
-            continue
-        lane_code = entry.get("lane_label")
-        lane_group = _lane_code_to_group(lane_code)
-        milestone = entry.get("milestone")
-        if milestone not in groups:
-            groups[milestone] = []
-        groups[milestone].append((repo, num, lane_code, lane_group))
-    return groups
-
-
-def _prepare_lane_metadata(layout: dict) -> dict[str, dict]:
-    """Build lane_code → {name, color, epic} lookup from layout."""
-    lane_meta: dict[str, dict] = {}
-    for lane in layout.get("lanes", []):
-        code = lane["code"]
-        lane_meta[code] = {
-            "name": lane.get("name", code.upper()),
-            "color": lane.get("color", code),
-            "epic": lane.get("epic"),
-        }
-    return lane_meta
-
-
-def _prepare_milestone_rows(  # noqa: C901
-    layout: dict,
-    gh_issues: dict,
-    primary_repo: str,
-    overrides: dict,
-) -> tuple[list[dict], dict[tuple[str, int], str], dict[str, dict]]:
-    """Prepare milestone-first row structure.
-
-    Returns (milestone_rows, lane_of, lane_meta) where:
-      - milestone_rows: list of {milestone, issues: [(repo, num, ...)]}
-      - lane_of: maps (repo, issue) → lane_code for cross-lane arrows
-      - lane_meta: lane_code → {name, color, epic}
-    """
-    lane_meta = _prepare_lane_metadata(layout)
-    groups = _group_issues_by_milestone(gh_issues, primary_repo)
-
-    # Also include issues from layout lanes that aren't in gh.json (render as missing)
-    seen: set[tuple[str, int]] = set()
-    for _ms, issues in groups.items():
-        for repo, num, _lane, _group in issues:
-            seen.add((repo, num))
-
-    for lane in layout.get("lanes", []):
-        lane_code = lane["code"]
-        lane_group = _lane_code_to_group(lane_code)
-        for item in lane.get("order", []):
-            if isinstance(item, dict):
-                repo = item["repo"]
-                num = item["issue"]
-            else:
-                repo = primary_repo
-                num = int(item)
-            if (repo, num) not in seen:
-                # Issue not in gh.json - add to "No milestone" group
-                # for missing placeholder rendering
-                if None not in groups:
-                    groups[None] = []
-                groups[None].append((repo, num, lane_code, lane_group))
-                seen.add((repo, num))
-
-    # Build lane_of mapping for cross-lane dependency arrows
-    lane_of: dict[tuple[str, int], str] = {}
-    for _ms, issues in groups.items():
-        for repo, num, lane_code, _lane_group in issues:
-            if lane_code:
-                lane_of[(repo, num)] = lane_code
-
-    # Sort issues within each milestone by lane group, then lane code, then issue
-    def sort_key(item: tuple[str, int, str | None, str | None]) -> tuple[int, int, int]:
-        _repo, num, lane, group = item
-        # Sort by group (A=0, B=1, ...), then lane code, then issue number
-        group_idx = ord(group) - ord("A") if group else 99
-        lane_idx = (
-            0
-            if lane is None
-            else ord(lane[0]) * 100 + (int(lane[1:]) if lane[1:].isdigit() else 0)
-        )
-        return (group_idx, lane_idx, num)
-
-    milestone_rows: list[dict] = []
-    for ms in MILESTONE_ORDER:
-        issues = groups.get(ms, [])
-        if not issues:
-            continue
-        sorted_issues = sorted(issues, key=sort_key)
-        milestone_rows.append(
-            {
-                "milestone": ms,
-                "issues": sorted_issues,
-            }
-        )
-
-    # Add any milestones not in MILESTONE_ORDER (e.g., legacy phases)
-    for ms in sorted(str(m) if m else "" for m in groups.keys()):
-        ms_key = None if ms == "" else ms
-        if ms_key not in MILESTONE_ORDER:
-            sorted_issues = sorted(groups[ms_key], key=sort_key)
-            milestone_rows.append(
-                {
-                    "milestone": ms,
-                    "issues": sorted_issues,
-                }
-            )
-
-    return milestone_rows, lane_of, lane_meta
-
-
-def _render_milestone_card(  # noqa: PLR0913
-    repo: str,
-    issue_num: int,
-    lane_code: str | None,
-    lane_meta: dict[str, dict],
-    lane_of: dict[tuple[str, int], str],
-    gh_issues: dict,
-    overrides: dict,
-    primary_repo: str,
-) -> str:
-    """Render a single card within a milestone row."""
-    gh_key = format_key(repo, issue_num)
-    gh_entry = gh_issues.get(gh_key)
-
-    # If issue is missing from gh.json, render a placeholder
-    if gh_entry is None:
-        return _render_missing_card(repo, issue_num)
-
-    ovr_key = gh_key
-    ovr = overrides.get(ovr_key, {})
-
-    status = derive_status(ovr, gh_entry, [], gh_issues, repo)
-    title = display_title(issue_num, ovr, gh_entry, None)
-    deps_html = render_deps(
-        CardContext(
-            repo=repo,
-            issue_num=issue_num,
-            lane_code=lane_code or "",
-            lane_of=lane_of,
-            ovr=ovr,
-            gh_entry=gh_entry,
-            extra_blocked_by=[],
-            extra_blocking=[],
-            gh_issues=gh_issues,
-            title_rules=None,
-            primary_repo=primary_repo,
-        )
-    )
-
-    # Repo badge for foreign repos (top right of card)
-    repo_badge_html = _render_repo_badge(repo, primary_repo)
-
-    card_id = f"card-{repo_slug(repo)}-{issue_num}"
-    top_inner = f'<span class="num">#{issue_num}</span>{repo_badge_html}'
-    return (
-        f'<div class="card {status}" id="{card_id}">'
-        f'<div class="top">{top_inner}</div>'
-        f'<div class="title">{title}</div>'
-        f'<div class="deps">{deps_html}</div>'
-        f"</div>"
-    )
-
-
-def render_milestone_row(  # noqa: C901, PLR0913
-    row: dict,
-    lane_meta: dict[str, dict],
-    lane_of: dict[tuple[str, int], str],
-    gh_issues: dict,
-    overrides: dict,
-    primary_repo: str,
-) -> str:
-    """Render one milestone row with header and cards in lane columns."""
-    milestone = row["milestone"]
-    issues = row["issues"]  # [(repo, num, lane_code, lane_group)]
-
-    if not issues:
-        return ""
-
-    # Header: "No milestone" or milestone title
-    if milestone is None:
-        header_html = (
-            '<div class="milestone-row-header" data-milestone="none">'
-            '<span class="ms-anchor">No milestone</span>'
-            '<span class="ms-hint">— independent lanes, standalone, closed</span>'
-            "</div>"
-        )
-    else:
-        header_html = (
-            f'<div class="milestone-row-header">'
-            f'<span class="ms-label">{escape(milestone)}</span>'
-            f"</div>"
-        )
-
-    # Group issues by lane_group (column), then by lane_code within each group
-    by_group: dict[str | None, dict[str | None, list[tuple[str, int]]]] = {}
-    for repo, num, lane_code, lane_group in issues:
-        if lane_group not in by_group:
-            by_group[lane_group] = {}
-        if lane_code not in by_group[lane_group]:
-            by_group[lane_group][lane_code] = []
-        by_group[lane_group][lane_code].append((repo, num))
-
-    # Render columns (lane groups A-I)
-    columns_html: list[str] = []
-    for group in ["A", "B", "C", "D", "E", "F", "G", "H", "I", None]:
-        if group not in by_group:
-            continue
-
-        group_meta = (
-            LANE_GROUP_META[group] if group else {"name": "Other", "color": "indep"}
-        )
-        by_lane = by_group[group]
-        lane_codes = sorted(by_lane.keys(), key=lambda x: (x or "zzz",))
-
-        # Column header (spans all sub-lanes if any)
-        col_header = (
-            f'<div class="lane-col-head" data-lane="{group_meta["color"]}">'
-            f'<span class="code">{group or "?"}</span>'
-            f'<span class="name">{escape(group_meta["name"])}</span></div>'
-        )
-
-        # If multiple sub-lanes, render as nested sub-columns
-        if len(lane_codes) > 1:
-            sub_cols_html: list[str] = []
-            for lane_code in lane_codes:
-                if lane_code is None:
-                    continue
-                lane_issues = by_lane[lane_code]
-                meta = lane_meta.get(
-                    lane_code, {"name": lane_code.upper(), "color": "indep"}
-                )
-
-                # Sub-column header
-                sub_header = (
-                    f'<div class="lane-sub-head" data-lane="{meta["color"]}">'
-                    f'<span class="code">{escape(lane_code.upper())}</span></div>'
-                )
-
-                # Render cards for this sub-lane
-                cards_html = [
-                    _render_milestone_card(
-                        repo,
-                        num,
-                        lane_code,
-                        lane_meta,
-                        lane_of,
-                        gh_issues,
-                        overrides,
-                        primary_repo,
-                    )
-                    for repo, num in lane_issues
-                ]
-
-                inner = "\n".join(cards_html)
-                sub_cols_html.append(
-                    f'<div class="lane-sub-col">\n'
-                    f"{sub_header}\n"
-                    f'<div class="lane-sub-cards">\n{inner}\n</div>\n'
-                    f"</div>"
-                )
-
-            sub_cols = "\n".join(sub_cols_html)
-            columns_html.append(
-                f'<div class="lane-col lane-col--split">\n'
-                f"{col_header}\n"
-                f'<div class="lane-sub-cols">\n{sub_cols}\n</div>\n'
-                f"</div>"
-            )
-        else:
-            # Single lane or no lane: render cards directly
-            lane_code = lane_codes[0] if lane_codes else None
-            lane_issues = by_lane.get(lane_code, [])
-
-            cards_html = [
-                _render_milestone_card(
-                    repo,
-                    num,
-                    lane_code,
-                    lane_meta,
-                    lane_of,
-                    gh_issues,
-                    overrides,
-                    primary_repo,
-                )
-                for repo, num in lane_issues
-            ]
-
-            inner = "\n".join(cards_html)
-            columns_html.append(
-                f'<div class="lane-col">\n'
-                f"{col_header}\n"
-                f'<div class="lane-col-cards">\n{inner}\n</div>\n'
-                f"</div>"
-            )
-
-    columns = "\n".join(columns_html)
-    return (
-        f'<div class="milestone-row">\n'
-        f"{header_html}\n"
-        f'<div class="milestone-cols">\n{columns}\n</div>\n'
-        f"</div>"
-    )
-
-
-def _prepare_render_data(
-    layout: dict,
-    gh_issues: dict,
-    primary_repo: str,
-    overrides: dict,
-) -> tuple[
-    list[dict],
-    list[tuple[str, int]],
-    list[tuple[str, int]],
-    dict[tuple[str, int], str],
-]:
-    """Derive lanes, resolve standalone/untriaged.
-
-    Returns (flat_lanes, standalone_order, untriaged, lane_of) where lane_of
-    maps (repo, issue) → lane code for every item placed in a lane — used by
-    downstream renderers for cross-lane dep arrows.
-    """
-    raw_lanes = layout["lanes"]
-    lanes = [derive_lane(lane, gh_issues, primary_repo) for lane in raw_lanes]
-    if is_auto_derived_standalone(layout):
-        standalone = {"order": derive_standalone_order(gh_issues, primary_repo)}
-    else:
-        standalone = layout["standalone"]
-
-    lane_of: dict[tuple[str, int], str] = {}
-    for lane in lanes:
-        for item in lane.get("order", []):
-            if isinstance(item, dict):
-                lane_of[(item["repo"], item["issue"])] = lane["code"]
-            else:
-                lane_of[(primary_repo, int(item))] = lane["code"]
-
-    flat_lanes = inject_spacers(
-        [flatten_lane(lane, overrides, True, gh_issues) for lane in lanes]
-    )
-
-    raw_standalone_order = standalone.get("order", [])
-    standalone_order: list[tuple[str, int]] = []
-    for item in raw_standalone_order:
-        if isinstance(item, dict):
-            standalone_order.append((item["repo"], item["issue"]))
-        else:
-            standalone_order.append((primary_repo, int(item)))
-
-    all_ordered: set[tuple[str, int]] = set(lane_of.keys()) | set(standalone_order)
-    epic_issues: set[tuple[str, int]] = {
-        (epic.get("repo", primary_repo), epic["issue"])
-        for lane in layout["lanes"]
-        if (epic := lane.get("epic"))
-    }
-    untriaged: list[tuple[str, int]] = sorted(
-        (
-            (entry.get("repo", primary_repo), entry["number"])
-            for _, entry in gh_issues.items()
-            if entry
-            and entry.get("lane_label") is not None
-            and not entry.get("hidden")
-            and (entry.get("repo", primary_repo), entry["number"]) not in all_ordered
-            and (entry.get("repo", primary_repo), entry["number"]) not in epic_issues
-            and not entry.get("standalone")
-            and "number" in entry
-        ),
-        key=lambda x: (x[0], x[1]),
-    )
-    return flat_lanes, standalone_order, untriaged, lane_of
 
 
 def build_html(layout: dict, gh_issues: dict) -> str:
@@ -1813,14 +108,22 @@ def build_html(layout: dict, gh_issues: dict) -> str:
 
     milestone_html = "\n\n".join(
         render_milestone_row(
-            row, lane_meta, lane_of, gh_issues, overrides, primary_repo
+            row,
+            lane_meta,
+            lane_of,
+            gh_issues,
+            overrides,
+            primary_repo,
+            derive_status=derive_status,
+            display_title=display_title,
+            render_deps=render_deps,
+            CardContext=CardContext,
+            _render_missing_card=_render_missing_card,
+            _render_repo_badge=_render_repo_badge,
         )
         for row in milestone_rows
     )
     cross_html = render_cross_deps(cross_deps)
-
-    # CSS — str.replace avoids .format() colliding with raw CSS braces
-    css = CSS_BASE.replace("{extra_vars}", "").replace("{extra_selectors}", "")
 
     escaped = _escape_meta(meta, primary_repo)
     title = escaped["title"]
@@ -1862,79 +165,24 @@ def build_html(layout: dict, gh_issues: dict) -> str:
         "B:#727</span> = cross-lane ref</span>"
     )
 
-    return f"""<!DOCTYPE html>
-<html lang="en" data-theme="dark">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta property="og:title" content="{title}">
-<meta property="og:description" content="{title}">
-<meta property="og:type" content="article">
-<meta property="og:url" content="https://forge.roxabi.dev/lyra/visuals/lyra-v2-dependency-graph.html">
-<meta property="og:image" content="https://forge.roxabi.dev/og-image.png">
-<meta property="og:image:width" content="1200">
-<meta property="og:image:height" content="630">
-<meta property="og:site_name" content="Roxabi Forge">
-<meta name="twitter:card" content="summary_large_image">
-<meta name="twitter:title" content="{title}">
-<meta name="twitter:description" content="{title}">
-<meta name="twitter:image" content="https://forge.roxabi.dev/og-image.png">
-<!-- diagram-meta:start -->
-<meta name="diagram:title"     content="{title}">
-<meta name="diagram:date"      content="{date}">
-<meta name="diagram:category"  content="{category}">
-<meta name="diagram:cat-label" content="{cat_label}">
-<meta name="diagram:color"     content="{color}">
-<meta name="diagram:badges"    content="latest">
-<meta name="diagram:issue"     content="{issue}">
-<!-- diagram-meta:end -->
-<title>{title}</title>
-<link rel="preconnect" href="https://fonts.googleapis.com">
-<link href="{fonts_url}" rel="stylesheet">
-<style>
-{css}
-</style>
-</head>
-<body>
-
-<header>
-  <div>
-    <h1>Lyra <span>v2</span> \u2014 Dependency Graph</h1>
-    <div class="subtitle">{subtitle}</div>
-  </div>
-  <div class="legend">
-    <span class="pill"><span class="dot done"></span> done</span>
-    <span class="pill"><span class="dot ready"></span> ready</span>
-    <span class="pill"><span class="dot blocked"></span> blocked</span>
-    <span class="pill"><span class="dot defer"></span> deferred</span>
-    {legend_epic}
-    {legend_xref}
-  </div>
-</header>
-
-<div class="milestones">
-
-{milestone_html}
-
-</div>
-
-<div class="cross-deps">
-  <h3>Cross-lane critical path (to M3 observability live)</h3>
-  <ul>
-{cross_html}
-  </ul>
-</div>
-
-<button class="theme-btn" id="theme-toggle">&#9681; light</button>
-
-<footer>
-  <p>{footer_line}</p>
-</footer>
-
-{THEME_SCRIPT}
-
-</body>
-</html>"""
+    template = _jinja_env.get_template("base.html")
+    return template.render(
+        title=title,
+        date=date,
+        issue=issue,
+        category=category,
+        cat_label=cat_label,
+        color=color,
+        fonts_url=fonts_url,
+        subtitle=subtitle,
+        legend_epic=legend_epic,
+        legend_xref=legend_xref,
+        footer_line=footer_line,
+        extra_vars="",
+        extra_selectors="",
+        milestone_html=milestone_html,
+        cross_html=cross_html,
+    )
 
 
 def run_build(
