+++ scripts/dep-graph/dep_graph/templates/base.html
@@ -1,0 +1,73 @@
+<!DOCTYPE html>
+<html lang="en" data-theme="dark">
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta property="og:title" content="{{ title }}">
+<meta property="og:description" content="{{ title }}">
+<meta property="og:type" content="article">
+<meta property="og:url" content="https://forge.roxabi.dev/lyra/visuals/lyra-v2-dependency-graph.html">
+<meta property="og:image" content="https://forge.roxabi.dev/og-image.png">
+<meta property="og:image:width" content="1200">
+<meta property="og:image:height" content="630">
+<meta property="og:site_name" content="Roxabi Forge">
+<meta name="twitter:card" content="summary_large_image">
+<meta name="twitter:title" content="{{ title }}">
+<meta name="twitter:description" content="{{ title }}">
+<meta name="twitter:image" content="https://forge.roxabi.dev/og-image.png">
+<!-- diagram-meta:start -->
+<meta name="diagram:title"     content="{{ title }}">
+<meta name="diagram:date"      content="{{ date }}">
+<meta name="diagram:category"  content="{{ category }}">
+<meta name="diagram:cat-label" content="{{ cat_label }}">
+<meta name="diagram:color"     content="{{ color }}">
+<meta name="diagram:badges"    content="latest">
+<meta name="diagram:issue"     content="{{ issue }}">
+<!-- diagram-meta:end -->
+<title>{{ title }}</title>
+<link rel="preconnect" href="https://fonts.googleapis.com">
+<link href="{{ fonts_url }}" rel="stylesheet">
+<style>
+{% include 'dep-graph.css' %}
+</style>
+</head>
+<body>
+
+<header>
+  <div>
+    <h1>Lyra <span>v2</span> — Dependency Graph</h1>
+    <div class="subtitle">{{ subtitle }}</div>
+  </div>
+  <div class="legend">
+    <span class="pill"><span class="dot done"></span> done</span>
+    <span class="pill"><span class="dot ready"></span> ready</span>
+    <span class="pill"><span class="dot blocked"></span> blocked</span>
+    <span class="pill"><span class="dot defer"></span> deferred</span>
+    {{ legend_epic }}
+    {{ legend_xref }}
+  </div>
+</header>
+
+<div class="milestones">
+
+{{ milestone_html }}
+
+</div>
+
+<div class="cross-deps">
+  <h3>Cross-lane critical path (to M3 observability live)</h3>
+  <ul>
+{{ cross_html }}
+  </ul>
+</div>
+
+<button class="theme-btn" id="theme-toggle">&#9681; light</button>
+
+<footer>
+  <p>{{ footer_line }}</p>
+</footer>
+
+{% include 'theme.js' %}
+
+</body>
+</html>
