+++ scripts/dep-graph/dep_graph/templates/card.html
@@ -1,0 +1,5 @@
+<div class="card {{ status }}" id="{{ card_id }}"{% if anchor_attr %}{{ anchor_attr }}{% endif %}>
+<div class="top">{{ top_inner }}</div>
+<div class="title">{{ title }}</div>
+<div class="deps">{{ deps_html }}</div>
+</div>
