+++ scripts/dep-graph/dep_graph/templates/lane_col.html
@@ -1,0 +1,8 @@
+<div class="lane-col">
+{{ col_header }}
+<div class="lane-col-cards">
+{% for card in cards %}
+{{ card }}
+{% endfor %}
+</div>
+</div>
