+++ scripts/dep-graph/dep_graph/templates/milestone_row.html
@@ -1,0 +1,8 @@
+<div class="milestone-row">
+{{ header_html }}
+<div class="milestone-cols">
+{% for col in columns %}
+{{ col }}
+{% endfor %}
+</div>
+</div>
